Adjust header width
Set the width of the header to be a little bigger than the main content
section. This will still need some adjusting, mainly header-row1,
because it shrink well. If I set a max-width, it pushes everything
together, but if I set a min-width or width, it stays that width even
when shrinking the window, and that looks bad. Will probably have
better ideas on how to fix this when I get to the section on responsive
design.